fix: overflow 버그 수정
- 로또 상금을 합치면서 int max값을 넘어가는 케이스 발생해서 int -> long 변경

+++ src/main/java/model/RewardCalculator.java
@@ -29,10 +29,10 @@ public class RewardCalculator {
 		return rankMatchingCounts.getOrDefault(rank, DEFAULT_COUNT);
 	}
 
-	public int sumReward() {
-		int sum = 0;
+	public long sumReward() {
+		long sum = 0;
 		for (Rank rank : rankMatchingCounts.keySet()) {
-			sum += rankMatchingCounts.get(rank) * rank.getReward();
+			sum += (long)rankMatchingCounts.get(rank) * rank.getReward();
 		}
 		return sum;
 	}
